Fix "On Testing" URL
The "On Testing" link in the first paragraph of the README doesn't
work, because the owner of that web page has moved it to a different
URL without the file extension, without setting up a redirect.

This commit fixes the URL so it goes to the page for which it is
intended.